+++ Sources/ProjectSpec/SpecOptions.swift
@@ -27,6 +27,7 @@ public struct SpecOptions: Equatable {
     public var groupSortPosition: GroupSortPosition
     public var generateEmptyDirectories: Bool
     public var findCarthageFrameworks: Bool
+    public var localPackagesGroup: String?
 
     public enum ValidationType: String {
         case missingConfigs
@@ -82,7 +83,8 @@ public struct SpecOptions: Equatable {
         transitivelyLinkDependencies: Bool = transitivelyLinkDependenciesDefault,
         groupSortPosition: GroupSortPosition = groupSortPositionDefault,
         generateEmptyDirectories: Bool = generateEmptyDirectoriesDefault,
-        findCarthageFrameworks: Bool = findCarthageFrameworksDefault
+        findCarthageFrameworks: Bool = findCarthageFrameworksDefault,
+        localPackagesGroup: String? = nil
     ) {
         self.minimumXcodeGenVersion = minimumXcodeGenVersion
         self.carthageBuildPath = carthageBuildPath
@@ -102,6 +104,7 @@ public struct SpecOptions: Equatable {
         self.groupSortPosition = groupSortPosition
         self.generateEmptyDirectories = generateEmptyDirectories
         self.findCarthageFrameworks = findCarthageFrameworks
+        self.localPackagesGroup = localPackagesGroup
     }
 }
 
@@ -129,6 +132,7 @@ extension SpecOptions: JSONObjectConvertible {
         groupSortPosition = jsonDictionary.json(atKeyPath: "groupSortPosition") ?? SpecOptions.groupSortPositionDefault
         generateEmptyDirectories = jsonDictionary.json(atKeyPath: "generateEmptyDirectories") ?? SpecOptions.generateEmptyDirectoriesDefault
         findCarthageFrameworks = jsonDictionary.json(atKeyPath: "findCarthageFrameworks") ?? SpecOptions.findCarthageFrameworksDefault
+        localPackagesGroup = jsonDictionary.json(atKeyPath: "localPackagesGroup")
     }
 }
 
@@ -149,6 +153,7 @@ extension SpecOptions: JSONEncodable {
             "indentWidth": indentWidth.flatMap { Int($0) },
             "tabWidth": tabWidth.flatMap { Int($0) },
             "defaultConfig": defaultConfig,
+            "localPackagesGroup": localPackagesGroup,
         ]
 
         if settingPresets != SpecOptions.settingPresetsDefault {