+++ Sources/ProjectSpec/SpecParsingError.swift
@@ -4,6 +4,7 @@ public enum SpecParsingError: Error, CustomStringConvertible {
     case unknownTargetType(String)
     case unknownTargetPlatform(String)
     case invalidDependency([String: Any])
+    case unknownPackageRequirement([String: Any])
     case invalidSourceBuildPhase(String)
     case invalidTargetReference(String)
     case invalidVersion(String)
@@ -22,6 +23,8 @@ public enum SpecParsingError: Error, CustomStringConvertible {
             return "Invalid Target Reference Syntax: \(targetReference)"
         case let .invalidVersion(version):
             return "Invalid version: \(version)"
+        case let .unknownPackageRequirement(package):
+            return "Unknown package requirement: \(package)"
         }
     }
 }